v0.6.0
The sheet was updated to have the latest functionalities that the system brings, like the option to roll from the critical injuries table, mark story points and keep score of how much cash the character has.

Besides that, a lot of small changes were made to improve the sheet readability and organization.